Use the Hilbert indice instead of cartesian indices
Neighbor and corner_neighbor now stores the hilbert indices of the
targeted patches.
smpi object now have a patch_count vector, listing the number of
patches owned by each mpi process has.
smpi now has a function hrank(h) which returns the rank of
the MPI process currently owning patch h.
The number of patch per MPI process is naively set to the
total number of patches divided by the number of MPI process.
A polymorphic version of the generalhilbertindex function in 2D
exists now. It does the same thing but it takes einit=dinit=0 by
default and does not return the value of e and d after last iteration.