Fix a major mathematical bug in LaserGaussian2D
The three-dimensional scaling of electric field amplitude with distance
from focus had been used instead of the two-dimensional one; this was
causing the laser amplitude to depend on the focal spot position as
well as a0, which it should not.
Now the same laser focused at different points reaches the same peak
intensity.