Add new 'vsym' kwarg to happi.
This kwarg being set to true indicates that the diagnostic returns data
which takes positive and negative values, and should be plotted using a
divergent colormap ('smileiD' by default) with limits centred on zero.
This allows animated plots of e.g. magnetic fields whose limits adjust
to the scale of fields present at each timestep, but retain a zero
midpoint.